[#4592] Make nuffic changes in one script
